+++ basics/counter/steel/api/src/consts.rs
@@ -1,0 +1,1 @@
+pub const COUNTER_SEEDS: &[u8] = b"counter"; // counter seeds
