+++ basics/counter/steel/api/src/state.rs
@@ -1,0 +1,15 @@
+use steel::*;
+
+#[repr(u8)]
+#[derive(Clone, Copy, Debug, Eq, PartialEq, IntoPrimitive, TryFromPrimitive)]
+pub enum CounterAccount {
+    Counter = 0,
+}
+
+account!(CounterAccount, Counter);
+// Counter
+#[repr(C, packed)]
+#[derive(Clone, Copy, Debug, PartialEq, Pod, Zeroable)]
+pub struct Counter {
+    pub count: u64,
+}
